feat: criando tipos para tratamento e fase
