--- src/types/VaccineTypes.ts
@@ -1,9 +1,0 @@
-export type VaccineType = {
-    id: number,
-    title: string,
-    observation: string,
-    status: string,
-    scheduledDate: string,
-    applicationDate: string,
-    phaseId?: number,
-}
